lib/test_sysctl: support testing of sysctl. boot parameter
Testing is done by a new parameter debug.test_sysctl.boot_int which defaults to
0 and it's expected that the tester passes a boot parameter that sets it to 1.
The test checks if it's set to 1. To distinguish true failure from parameter
not being set, the test checks /proc/cmdline for the expected parameter, and
whether test_sysctl is built-in and not a module.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>

Backport [PATCH v3 5/5]
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>